moved user routes into separate file
