+++ src/server.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { mockUsers } from "./mock-data/users";
+import userRouter from "./routes/userRoutes";
 
 const PORT = process.env.PORT || 8080;
 
@@ -8,45 +8,11 @@ const app = express();
 /* -------------------------------------------------------------------------- */
 /*                                   Routes                                   */
 /* -------------------------------------------------------------------------- */
+app.use("/api/users", userRouter);
 
 app.get("/", (req, res) => {
   res.send("Hello world!");
 });
-
-app.get("/api/users", (req, res) => {
-  // check for a "limit" query parameter
-
-  const queryLimit = req.query.limit;
-
-  if (typeof queryLimit === "undefined") {
-    res.status(200).json(mockUsers);
-    return;
-  }
-
-  const limit = parseInt(queryLimit as string);
-  if (limit < 0 || isNaN(limit)) {
-    res.status(400).send("Limit should be a positive number");
-    return;
-  }
-
-  const slicedUsers = mockUsers.slice(0, limit);
-
-  res.status(200).json(slicedUsers);
-});
-
-app.get("/api/users/:id", (req, res) => {
-  const userId = parseInt(req.params.id);
-  const user = mockUsers.find((user) => user.id === userId);
-
-  if (!user) {
-    res.status(404).send("User not found");
-    return;
-  }
-
-  res.status(200).json(user);
-});
-
-app.get;
 
 /* ------------------------------ Start server ------------------------------ */
 app.listen(PORT, () => {
